pyproject.toml
[tool.poetry]
name = "e-commerceproject"
version = "0.1.0"
description = "E-Commerce project for lectrue ECC431 ( E-Commerce )"
authors = ["Can İlgu <canilguu@gmail.com>"]
license = "Apache License Version 2.0"
readme = "README.md"
packages = [{include = "e_commerceproject"}]

[tool.poetry.dependencies]
python = "^3.8 || ^3.10"
fastapi = "0.87.0"
aiomysql = "0.1.1"
bcrypt = "4.0.1"
celery = "5.2.7"
redis = "^4.3.4"
PyYAML = "^6.0"
aioredis = "^2.0.1"
tortoise-orm = {extras = ["asyncmy"], version = "^0.19.2"}
httpx = "^0.23.0"
Jinja2 = "^3.1.2"
pytest-asyncio = "^0.20.2"
jose = "^1.0.0"
passlib = "^1.7.4"
boto3 = "^1.26.16"
loguru = "^0.6.0"
aerich = "^0.7.1"
fastapi-mail = "^1.2.2"
python-jose = "^3.3.0"
cryptography = "^38.0.4"
uvicorn = "^0.20.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"
flake8 = "^5.0.4"


[tool.aerich]
tortoise_orm = "core.settings.TORTOISE_ORM"
location = "./database/migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
